requirements.txt
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
arxiv==2.1.0
autogenstudio==0.0.31a0
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
diskcache==5.6.3
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.109.0
feedparser==6.0.10
FLAML==2.1.1
fonttools==4.47.2
fpdf==1.7.2
frozendict==2.4.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-resources==6.1.1
joblib==1.3.2
kiwisolver==1.4.5
lxml==5.1.0
matplotlib==3.8.2
multitasking==0.0.11
numpy==1.26.3
openai==1.8.0
packaging==23.2
pandas==2.2.0
peewee==3.17.0
pillow==10.2.0
pyautogen==0.2.7
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
regex==2023.12.25
requests==2.31.0
scikit-learn==1.4.0
scipy==1.11.4
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.35.1
tavily-python==0.3.0
termcolor==2.4.0
threadpoolctl==3.2.0
tiktoken==0.5.1
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.26.0
webencodings==0.5.1
yfinance==0.2.35
zipp==3.17.0
